Revert "formatting fix"
This reverts commit d5bda785d51ff7647b3df765a367fa98422f0679.

This wasn't an issue at all,
the real issue seems to be that referring to an absolute path is broken
instead.

i.e. referring just by "commonpins.md" results to the link to be
replaced to the HTML representation "commonpins.html".

But doing it like "/ip/commonpins.md" results to the link to refer
to the raw markdown document instead.